+++ CHANGELOG.md
@@ -200,3 +200,10 @@
 
 ### v1.0 /8/28/19
 * added john the ripper and snort to lib_install.sh
+
+### v1.0 /8/31/19
+* added tshark, netcat, unicornscan, fierce, openvas, nikto, wpscan
+* added fluxion necessary libs to lib_install: mawk, curl, dhcpd, hostapd, lighttpd, mdk3, php-cgi, pyrit, unizip, xterm, openssl, rfkill, strings
+* added fluxion to network_crack.py
+* added fluxion install to lib_install.sh
+* added list of lib_install.sh installation candidates: LIB_REGISTER.md